+++ spec/forms/conversion/edit_project_form_spec.rb
@@ -1,78 +1,112 @@
 require "rails_helper"
 
 RSpec.describe Conversion::EditProjectForm, type: :model do
-  let(:project) { build(:conversion_project, assigned_to: user) }
+  let(:project) { build(:conversion_project, assigned_to: user, advisory_board_date: Date.new(2023, 12, 1)) }
   let(:user) { build(:user, :caseworker) }
+
+  subject { Conversion::EditProjectForm.new_from_project(project) }
 
   before do
     mock_successful_api_responses(urn: any_args, ukprn: any_args)
   end
 
-  describe "#save" do
-    describe "Incoming trust UKPRN" do
+  describe "#update" do
+    describe "the incoming trust UKPRN" do
       it "can be changed" do
-        form = Conversion::EditProjectForm.new(project: project, args: {incoming_trust_ukprn: "12345678"})
-        form.save
+        updated_params = {incoming_trust_ukprn: "12345678"}
+
+        subject.update(updated_params)
 
         expect(project.incoming_trust_ukprn).to eql 12345678
       end
 
       it "cannot be invalid" do
-        form = Conversion::EditProjectForm.new(project: project, args: {incoming_trust_ukprn: "2461810"})
+        updated_params = {incoming_trust_ukprn: "2461810"}
 
-        expect(form.valid?).to be false
+        expect(subject.update(updated_params)).to be false
         expect(project.incoming_trust_ukprn).to eql 10061021
       end
 
       it "the trust must exist" do
         mock_trust_not_found(ukprn: 12345678)
 
-        form = Conversion::EditProjectForm.new(project: project, args: {incoming_trust_ukprn: "12345678"})
+        updated_params = {incoming_trust_ukprn: "12345678"}
 
-        expect(form.valid?).to be false
+        expect(subject.update(updated_params)).to be false
         expect(project.incoming_trust_ukprn).to eql 10061021
+      end
+    end
+
+    describe "the advisory board date" do
+      it "can be changed" do
+        updated_params = {
+          "advisory_board_date(3i)" => "1",
+          "advisory_board_date(2i)" => "1",
+          "advisory_board_date(1i)" => "2024"
+        }
+
+        subject.update(updated_params)
+
+        expect(project.advisory_board_date).to eql Date.new(2024, 1, 1)
+      end
+
+      it "cannot be invalid" do
+        updated_params = {
+          "advisory_board_date(3i)" => "32",
+          "advisory_board_date(2i)" => "1",
+          "advisory_board_date(1i)" => "2024"
+        }
+
+        expect(subject.update(updated_params)).to be false
+        expect(project.advisory_board_date).to eql project.advisory_board_date
       end
     end
 
     describe "Establishment SharePoint link" do
       it "can be changed" do
-        form = Conversion::EditProjectForm.new(project: project, args: {establishment_sharepoint_link: "https://educationgovuk-my.sharepoint.com/establishment-folder-updated-link"})
-        form.save
+        updated_params = {establishment_sharepoint_link: "https://educationgovuk-my.sharepoint.com/establishment-folder-updated-link"}
+
+        subject.update(updated_params)
 
         expect(project.establishment_sharepoint_link).to eql("https://educationgovuk-my.sharepoint.com/establishment-folder-updated-link")
       end
 
       it "cannot be invalid" do
-        form = Conversion::EditProjectForm.new(project: project, args: {establishment_sharepoint_link: "https://some-other-domain.com/establishment-folder-updated-link"})
+        updated_params = {establishment_sharepoint_link: "https://some-other-domain.com/establishment-folder-updated-link"}
 
-        expect(form.valid?).to be false
+        expect(subject.update(updated_params)).to be false
+        expect(project.establishment_sharepoint_link).to eql("https://educationgovuk-my.sharepoint.com/establishment-folder")
       end
 
       it "cannot be empty" do
-        form = Conversion::EditProjectForm.new(project: project, args: {establishment_sharepoint_link: ""})
+        updated_params = {establishment_sharepoint_link: ""}
 
-        expect(form.valid?).to be false
+        expect(subject.update(updated_params)).to be false
+        expect(project.establishment_sharepoint_link).to eql("https://educationgovuk-my.sharepoint.com/establishment-folder")
       end
     end
 
     describe "Incoming trust SharePoint link" do
       it "can be changed" do
-        form = Conversion::EditProjectForm.new(project: project, args: {incoming_trust_sharepoint_link: "https://educationgovuk-my.sharepoint.com/incoming-trust-folder-updated-link"})
-        form.save
+        updated_params = {incoming_trust_sharepoint_link: "https://educationgovuk-my.sharepoint.com/incoming-trust-folder-updated-link"}
+
+        subject.update(updated_params)
 
         expect(project.incoming_trust_sharepoint_link).to eql("https://educationgovuk-my.sharepoint.com/incoming-trust-folder-updated-link")
       end
 
       it "cannot be invalid" do
-        form = Conversion::EditProjectForm.new(project: project, args: {incoming_trust_sharepoint_link: "https://some-other-domain.com/incoming-trust-folder-updated-link"})
+        updated_params = {incoming_trust_sharepoint_link: "https://some-other-domain.com/incoming-trust-folder-updated-link"}
 
-        expect(form.valid?).to be false
+        expect(subject.update(updated_params)).to be false
+        expect(project.incoming_trust_sharepoint_link).to eql("https://educationgovuk-my.sharepoint.com/incoming-trust-folder")
       end
 
       it "cannot be empty" do
-        form = Conversion::EditProjectForm.new(project: project, args: {incoming_trust_sharepoint_link: ""})
+        updated_params = {incoming_trust_sharepoint_link: ""}
 
-        expect(form.valid?).to be false
+        expect(subject.update(updated_params)).to be false
+        expect(project.incoming_trust_sharepoint_link).to eql("https://educationgovuk-my.sharepoint.com/incoming-trust-folder")
       end
     end
   end