+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   export when available
 - users can now change the incoming trust UKPRN on a conversion project, this
   allows mistakes to be corrected.
+- users can now change the advisory board date and any conditions on a
+  conversion project.
 
 ## [Release-62][release-62]